+++ Assets/LDtkUnity/Runtime/Tools/SceneDrawer/LDtkSceneDrawerBase.cs
@@ -7,37 +7,21 @@ namespace LDtkUnity
     public abstract class LDtkSceneDrawerBase
     {
         protected Transform Transform;
-        protected Component Source;
+        protected string Identifier;
+        protected LDtkFields Fields;
         protected EditorDisplayMode Mode;
         protected float GridSize;
         
-        private string _fieldName;
         
-        public void SupplyReferences(Component source, string fieldName, EditorDisplayMode mode, float gridSize)
+        public void SupplyReferences(LDtkFields fields, string identifier, EditorDisplayMode mode, float gridSize)
         {
-            Source = source;
-            Transform = source.transform;
-            _fieldName = fieldName;
+            Fields = fields;
+            Transform = fields.transform;
+            Identifier = identifier;
             Mode = mode;
             GridSize = gridSize;
         }
-        
-        protected FieldInfo GetFieldInfo()
-        {
-            if (Source == null || string.IsNullOrEmpty(_fieldName))
-            {
-                Debug.LogError("GetField Error");
-                return null;
-            }
 
-            Type type = Source.GetType();
-                
-            FieldInfo fieldInfo = type.GetField(_fieldName);
-            return fieldInfo;
-        }
-        
         public abstract void Draw();
-
-
     }
 }